Implement pass and like for discover
