+++ app/src/main/res/layout/fragment_user_detail.xml
