+++ app/src/main/res/values/dimens.xml
@@ -4,4 +4,5 @@
     <dimen name="card_elevation">4dp</dimen>
     <dimen name="card_margin">23dp</dimen>
     <dimen name="image_view_height">550dp</dimen>
+    <dimen name="text_size">10dp</dimen>
 </resources>
